backend/requirements.txt
fastapi==0.95.1
langchain==0.0.147
python-dotenv==1.0.0
Requests==2.28.2
